complaint_system/requirements.txt
flask
flask-mail
mysql-connector-python
werkzeug
